dry-ing up SIWE messages
